requirements.txt
anomalib
torch
torchvision
torchaudio
lightning
openvino
onnx
onnxruntime
open-clip-torch
opencv-python
Pillow
scikit-image
numpy
pandas
scikit-learn
matplotlib
seaborn
plotly
PyYAML
tqdm
click
requests
tensorboard
pytest
black
flake8
scipy
huggingface-hub